+++ src/core/messaging/notifications/notifications.ts
@@ -5,8 +5,6 @@ import { notificationHandler } from '../handlers/notification';
 import { store } from '../../../redux/config';
 import { updateAddressMonitorTokens } from '../../address-monitor';
 import { themes } from '../../../navigation/navigation';
-import { HttpClient } from '../../utils/http-client';
-import CONFIG from '../../../config';
 
 // this file is in this format for testing purposes
 export class NotificationService {
@@ -71,28 +69,8 @@ export class NotificationService {
         this.onNotificationOpenedListener = firebase
             .notifications()
             .onNotificationOpened(notificationOpen => {
-                const notifData = (notificationOpen.notification as any).data;
-                notificationHandler(notifData, false);
-
-                // TODO: check this
-                // Also trigger when app is closed and the notif is opened
-                this.markSeenNotification(notifData);
+                notificationHandler((notificationOpen.notification as any).data, false);
             });
-    }
-
-    private async markSeenNotification(notifData: any) {
-        try {
-            const parsedData = JSON.parse(notifData.data);
-
-            if (parsedData?.notification?._id) {
-                const http = new HttpClient(CONFIG.notificationCenter.markSeenUrl);
-                await http.post('', {
-                    notifIds: [parsedData.notification._id]
-                });
-            }
-        } catch {
-            //
-        }
     }
 
     public removeListeners() {
